+++ NEWS.md
@@ -7,6 +7,8 @@
   being restricted to a single number.
 - `cumulative_cutoff()` parameter `cutoff` now accepts a numeric vector, instead
   of being restricted to a single number.
+- `cumulative_interval()` parameter `interval` now accepts a list of numeric
+  vectors, instead of being restricted to a single vector.
 - The decay functions can now take numeric vectors as input, instead of being
   restricted to a single number (in the case of `decay_stepped()`, both `steps`
   and `weights` can take a `list` of `numeric` vectors as input, instead of